os/bluestore/AllocatorBase.h: cast myTraits.base for p2aligned on macOS
p2aligned<T>(T,T) needs identical arg types; offset/len are uint64_t but
myTraits.base is size_t (same width, distinct type on Darwin), so deduction
fails. Cast to uint64_t at all three asserts.

+++ src/os/bluestore/AllocatorBase.h
@@ -209,8 +209,8 @@ protected:
         return false;
       }
       bool ret = false;
-      ceph_assert(p2aligned(offset, myTraits.base));
-      ceph_assert(p2aligned(len, myTraits.base));
+      ceph_assert(p2aligned(offset, static_cast<uint64_t>(myTraits.base)));
+      ceph_assert(p2aligned(len, static_cast<uint64_t>(myTraits.base)));
       auto idx = myTraits._get_bucket(len);
       if (idx < buckets.size())
         ret = buckets[idx].try_put(offset);
@@ -223,7 +223,7 @@ protected:
       }
       bool ret = false;
       ceph_assert(offset);
-      ceph_assert(p2aligned(len, myTraits.base));
+      ceph_assert(p2aligned(len, static_cast<uint64_t>(myTraits.base)));
       size_t idx = len >> myTraits.base_bits;
       if (idx < buckets.size()) {
         ret = buckets[idx].try_get(*offset);
